src/ml_templates/state_evolution/train_with_checkpoints/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "train-with-checkpoints"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Wesley Holt", email = "wesleyjholt@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "equinox",
  "optax",
  "torch",
  "torchdata",
  "orbax-checkpoint",
  "pyyaml"
]

[project.urls]
Documentation = "https://github.com/Wesley Holt/train-with-checkpoints#readme"
Issues = "https://github.com/Wesley Holt/train-with-checkpoints/issues"
Source = "https://github.com/Wesley Holt/train-with-checkpoints"

[tool.hatch.version]
path = "src/train_with_checkpoints/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0"
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/train_with_checkpoints tests}"

[tool.coverage.run]
source_pkgs = ["train_with_checkpoints", "tests"]
branch = true
parallel = true
omit = [
  "src/train_with_checkpoints/__about__.py",
]

[tool.coverage.paths]
train_with_checkpoints = ["src/train_with_checkpoints", "*/train-with-checkpoints/src/train_with_checkpoints"]
tests = ["tests", "*/train-with-checkpoints/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
